+++ InscryptionAPI/Pelts/PeltManager.cs
@@ -62,14 +62,15 @@ public static class PeltManager
 
     internal static int[] BasePeltPrices
     {
-        get
+        get // base w/o modifiers: 2, 4, 7
         {
-            BuyPeltsSequencer instance = SpecialNodeHandler.Instance.buyPeltsSequencer;
+            int expensivePeltsMult = AscensionSaveData.Data.ChallengeIsActive(AscensionChallenge.ExpensivePelts) ? 2 : 1;
+            int defeatedTrapperMult = SpecialNodeHandler.Instance.buyPeltsSequencer.TrapperBossDefeated ? 2 : 1;
             return new int[]
             {
-                2 / (instance.TrapperBossDefeated ? 2 : 1) * (AscensionSaveData.Data.ChallengeIsActive(AscensionChallenge.ExpensivePelts) ? 2 : 1),
-                (4 + RunState.CurrentRegionTier) / (instance.TrapperBossDefeated ? 2 : 1) * (AscensionSaveData.Data.ChallengeIsActive(AscensionChallenge.ExpensivePelts) ? 2 : 1),
-                Mathf.Min(20, (7 + RunState.CurrentRegionTier * 2) / (instance.TrapperBossDefeated ? 2 : 1) * (AscensionSaveData.Data.ChallengeIsActive(AscensionChallenge.ExpensivePelts) ? 2 : 1))
+                2 / defeatedTrapperMult * expensivePeltsMult,
+                (4 + RunState.CurrentRegionTier) / defeatedTrapperMult * expensivePeltsMult,
+                Mathf.Min(20, (7 + RunState.CurrentRegionTier * 2) / defeatedTrapperMult * expensivePeltsMult)
             };
         }
     }
@@ -125,13 +126,23 @@ public static class PeltManager
     /// </summary>
     /// <param name="pluginGuid">GUID of the mod adding this pelt.</param>
     /// <param name="peltCardInfo">The CardInfo for the actual pelt card.</param>
+    /// <param name="getCardChoices">The list of possible cards the Trader will offer for this pelt.</param>
     /// <param name="baseBuyPrice">The starting price of this pelt when buying from the Trapper.</param>
     /// <param name="extraAbilitiesToAdd">The number of extra sigils card choices will have when trading this pelt to the Trader.</param>
     /// <param name="choicesOfferedByTrader">How many cards to offer the player when trading the pelt.</param>
-    /// <param name="getCardChoices">The list of possible cards the Trader will offer for this pelt.</param>
     /// <returns>The newly created CustomPeltData so a chain can continue.</returns>
+    public static PeltData New(string pluginGuid, CardInfo peltCardInfo, Func<List<CardInfo>> getCardChoices, int baseBuyPrice, int extraAbilitiesToAdd = 0, int choicesOfferedByTrader = 8)
+    {
+        return New(pluginGuid, peltCardInfo, baseBuyPrice, extraAbilitiesToAdd, choicesOfferedByTrader, getCardChoices);
+    }
+
     public static PeltData New(string pluginGuid, CardInfo peltCardInfo, int baseBuyPrice, int extraAbilitiesToAdd, int choicesOfferedByTrader, Func<List<CardInfo>> getCardChoices)
     {
+        if (getCardChoices == null)
+        {
+            throw new ArgumentNullException("CardChoices function cannot be null!");
+        }
+
         PeltData peltData = new()
         {
             pluginGuid = pluginGuid,
@@ -168,12 +179,7 @@ public static class PeltManager
     public static List<PeltData> AllPeltsAvailableAtTrader()
     {
         List<PeltData> peltNames = new();
-        foreach (PeltData data in AllPelts())
-        {
-            if (data.isSoldByTrapper)
-                peltNames.Add(data);
-        }
-
+        peltNames.AddRange(AllPelts().Where(x => x.isSoldByTrapper));
         return peltNames;
     }